Token authentication is now an optional feature
