+++ .rustme/docs.md
@@ -83,6 +83,8 @@ bonsaidb = { version = "*", features = "full" }
 - `cli`: Enables the `bonsaidb` executable.
 - `password-hashing`: Enables the ability to use password authentication using
   Argon2 via `AnyConnection`.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
 
 All other feature flags, listed below, affect each crate individually, but can
 be safely combined.
@@ -107,6 +109,8 @@ All Cargo features that affect local databases:
 - `instrument`: Enables instrumenting with `tracing`.
 - `password-hashing`: Enables the ability to use password authentication
   using Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
 
 ### BonsaiDb server
 
@@ -130,6 +134,8 @@ All Cargo features that affect networked servers:
 - `websockets`: Enables `WebSocket` support.
 - `password-hashing`: Enables the ability to use password authentication
   using Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
 
 ### Client for accessing a BonsaiDb server
 
@@ -148,3 +154,5 @@ All Cargo features that affect networked clients:
 - `websockets`: Enables `WebSocket` support for `bonsaidb-client`.
 - `password-hashing`: Enables the ability to use password authentication
   using Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
