+++ crates/bonsaidb-client/README.md
@@ -49,6 +49,8 @@ By default, the `full` feature is enabled.
 - `websockets`: Enables `WebSocket` support for `bonsaidb-client`.
 - `password-hashing`: Enables the ability to use password authentication
   using Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
 - `tracing`: Enables `tracing` annotations on some functions and dependencies.
 
 ## Open-source Licenses
