+++ crates/bonsaidb-local/local-feature-flags.md
@@ -11,3 +11,5 @@ By default, the `full` feature is enabled.
 - `multiuser`: Enables multi-user support.
 - `password-hashing`: Enables the ability to use password authentication using
   Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
