+++ crates/bonsaidb-server/Cargo.toml
@@ -38,6 +38,10 @@ password-hashing = [
     "bonsaidb-local/password-hashing",
     "bonsaidb-core/password-hashing",
 ]
+token-authentication = [
+    "bonsaidb-core/token-authentication",
+    "bonsaidb-local/token-authentication",
+]
 compression = ["bonsaidb-local/compression"]
 
 included-from-omnibus = []
