+++ crates/bonsaidb-server/README.md
@@ -36,6 +36,8 @@ By default, the `full` feature is enabled.
 - `websockets`: Enables `WebSocket` support.
 - `password-hashing`: Enables the ability to use password authentication
   using Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
 
 ## Open-source Licenses
