+++ crates/bonsaidb-server/server-feature-flags.md
@@ -12,3 +12,5 @@ By default, the `full` feature is enabled.
 - `websockets`: Enables `WebSocket` support.
 - `password-hashing`: Enables the ability to use password authentication
   using Argon2.
+- `token-authentication`: Enables the ability to authenticate using
+  authentication tokens, which are similar to API keys.
